+++ examples/ai-core/src/stream-object/openai-on-finish.ts
@@ -21,11 +21,22 @@ async function main() {
     }),
     prompt:
       'Generate 3 character descriptions for a fantasy role playing game.',
+    onFinish({ usage, object, error }) {
+      console.log();
+      console.log('onFinish');
+      console.log('Token usage:', usage);
+
+      // handle type validation failure (when the object does not match the schema):
+      if (object === undefined) {
+        console.error('Error:', error);
+      } else {
+        console.log('Final object:', JSON.stringify(object, null, 2));
+      }
+    },
   });
 
+  // consume the partialObjectStream:
   for await (const partialObject of result.partialObjectStream) {
-    console.clear();
-    console.log(partialObject);
   }
 }