+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/openai
 
+## 0.0.24
+
+### Patch Changes
+
+- 0e78960c: fix (@ai-sdk/openai): make function name and arguments nullish
+
+## 0.0.23
+
+### Patch Changes
+
+- a68fe74a: fix (@ai-sdk/openai): allow null tool_calls value.
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [102ca22f]
+  - @ai-sdk/provider@0.0.10
+  - @ai-sdk/provider-utils@0.0.13
+
 ## 0.0.21
 
 ### Patch Changes